Move recordings to different files
