+++ Core/DependencyRuleEngine/Feedback/Recorder/Recordings/ContextRecording.cs
@@ -1,0 +1,7 @@
+namespace YADA.Core.DependencyRuleEngine.Feedback
+{
+    internal class ContextRecording 
+    {
+        public string Context { get; set; }
+    }
+}
