+++ Core/DependencyRuleEngine/Feedback/Recorder/Recordings/InfoRecording.cs
@@ -1,0 +1,7 @@
+namespace YADA.Core.DependencyRuleEngine.Feedback
+{
+    internal class InfoRecording 
+    {
+        public string Info { get; set; }
+    }
+}
